+++ src/backend/profileFiltersHelper.ts
@@ -491,6 +491,8 @@ const filterAutorisationCapaciteEJ = (autorisationsEtCapacites: any, profil: any
   const filtredAutorisationCapacite = {
     autorisationsActivités:
       profil.autorisationsActivités === "ok" ? autorisationsEtCapacites.autorisationsActivités : { dateMiseÀJourSource: "", autorisations: [] },
+    autorisationsMédicoSocial:
+      profil.autorisationsActivités === "ok" ? autorisationsEtCapacites.autorisationsMédicoSocial : { dateMiseÀJourSource: "", autorisations: [] },
     autorisationsAmmSanitaire:
       profil.autorisationsActivités === "ok" ? autorisationsEtCapacites.autorisationsAmmSanitaire : { dateMiseÀJourSource: "", autorisations: [] },
     autresActivités: profil.autresActivités === "ok" ? autorisationsEtCapacites.autresActivités : { dateMiseÀJourSource: "", autorisations: [] },